ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
